Made changes to balloon spawns, as they were spawning out of bounds.
Modified Gamemanger for timer ending game, score management and health and showing word to rhyme with.

Added ui for what to rhyme with.

Changed Object spawner to use the prefab and 0'd the coords for the prefab.

Fixed issue with stopping at the waypoint 1 after the one to stop at and all stopping points reactivate.

Fixed pause menu scaling.